Fixed example
